+++ docs/rts_overlay.js
@@ -74,6 +74,20 @@ const EXTERNAL_BO_WEBSITES = {
       "Click on 'Open in RTS Overlay'.",
     ],
   ],
+  sc2: [
+    [
+      'RTS Builds',
+      'https://craftysalamander.github.io/rtsbuilds?gameId=sc2',
+      "Click on 'Open in RTS Overlay'.",
+    ],
+  ],
+  wc3: [
+    [
+      'RTS Builds',
+      'https://craftysalamander.github.io/rtsbuilds?gameId=wc3',
+      "Click on 'Open in RTS Overlay'.",
+    ],
+  ],
 };
 
 // Fields of the faction name: player and (optionally) opponent
@@ -1942,6 +1956,10 @@ function checkBuildOrderKeyValues(buildOrder, keyCondition = null) {
 
   // Loop  on the key conditions
   for (const [key, target] of Object.entries(keyCondition)) {
+    if (['all', 'All'].includes(target)) {
+      // all are valid
+      continue;
+    }
     if (key in buildOrder) {
       const dataCheck = buildOrder[key];
       // Any build order data value is valid